Sign up and sign in with email and password complet
user can now sign in and asign up with email and pasword. They will not allowed access either login and register if they are logged in. they need to log out to log back. Upon logging in or registering, the user will be brought back to mainactivity.

TODO

Add user to users table in RTDB