Action Bar updated
Action bar now has a menu of items which contain the logout button for the mainactivity.
Checks if the user is logged. if they are not it will open the login activity and and close the main activity
The colour theme was updated.
Logout button removed from mainactivity in favour of the menu logout solution