Create and join farm added and changes to databaseManager
Create and join farm are called if there is the users farm in the users table is none. In this state it currently does all the time as there is no farm table in the db yet.

User now contains the farm id and the main activty queries the db to check if the user is in a farm or not.

adding a user to the database has now changed as the user id and farm id is part of the user object. instead of the object and the user id being passed, the user id is added to the object and only the object is passed to the db manager.