Added create farm and join farm features
Allows users to create a farm and to join farms. create makes a new table and then adds the user to the list of users of that farm. it also adds the farm name to user in the database. Validation for inputs and database querying is done. checks farm exists before making it.

Join farm checks that the user farm entered exists and then checks if the pass codes match.

Database manager expanded to incclude addFarmToDatabase, AddFarmNameToUserAndUserToFarm, getDatabaseRefForUserId