Added ManagerTimeTable Activity
In test phase.

Added new custom listview with listview adapter.

Timeslot layout made for each row of the listview. Timeslot class created with constructor and getter and setters.

Timetable Listview Adapter class creates a custom listview row that displays the time, name and description.

manager timetable currently hardcodes a timetable that is pasted to check the adpater works. The next phase will be make a function in the timetable services that creates a template of data for each day that can then have tasks written to it.