ManagerTime update
Converted timetable view to manager timetable view. Allows managers to set users timetable by selecting a a date, start time, end time and make a db call to retrieve all the users that the task can be assigned to.

MainActivity edited to add check incase  there is no current firebase user. In the case of a fresh install, a check needs to be made to check if the instance was null. Previously it checked if the user was null. This caused an error as on a fresh install, Firebase Auth would be null and therefore Auth.getCurrentUser would give the error.

Change to Database manager when adding a user to the list of users in a farm. The users name is gotten from the users table, combined to be one string. This is then used as the value in the usersinfarm sub table in the farm table while the key remains the user id.