Task now added to users timetable.
Task object created which holds information about the task. Includes extra information such as colour and complete status. These features can be added later at the end of the project once the requirements are met.

When the manager enters the details and selects the staff member to assign the task to, the time table is then added to that users timetable in usersInFarm table inside the farm table. This way each user can have their own personalized timetable. Each task in the timetable table has a unique id to identify it.