Manager Options for order board
Manager can click to open a view which acts as a form page. There are two pop up windows that become visible when the user presses the relevant button. The inputs on the original form are disabled when the pop up is open and are re enabled when the window is closed.

The user will see a drop down list of customers and a drop down list of products. They can also pick a date and then enter a quantity. If a customer is not selected because there were none in the database, a message will remind them that they need to add one. the same applies for customers.

When the user changes customer in the drop down list, the products list also changes.

When adding a customer, the user only needs to enter the name of the customer hit enter. when adding a product to a customer, the user can select the existing customers from the dropdown menu.

At this commit, customers and products can be added and the drop down menus can find the data they need. If a product or customer exists, the user will be told.

TODO
Add orders based off date, customer, product and quantity.
Make a list adapter that can display header and products. Can also be clickable to change status from an X to a tick.