Orders can now be added
Orders can now be added based off what customer is selected and the product.

The order is added into the customer directory and the product directory.
/farm_table/farmName/customers/customerName/productName/orders/orderId .

If an order is put in that already exists on that date, it will be replaced and the user will be informed that it has been replaced rather than being added as new.