OrderBoard Displaying Template
OrderBoard call testOrderboard() which populates the order board. It uses a custome List adapter to display 3 columns. If the quantity column is set as null, The adapter will interpret this as being a customer name header and only set the name and change the background colours.

TODO
Read data from the database and populate the table