Order board can now read orders form the database
The orders board can read using updateOrderBoardWithDate()

It creates a list of order rows andloops through customers, then through porducts and then through orders and adds all the orders that match the date provided. It then passes on that object to the the OrdersBoardListAdapter() which formats and displays it the user.