Updated Ordersboard Listview and added onclick to change the status
The Orderboard row layout was modified to include all the attribute of the Order object. The one not needed are hidden but it allows all the data of the order to be stored for the onclick listener. the OrdersBoardListAdapter was also modified to allow this to display the new list object.

The onClick listener is set in updateOrderBoardAndListViewOnClickListenerWithDate(). This is called when the order board is first loaded. The listener is set each time the order board is updated to reset it with the new values. Once a row is clicked, it is filtered to make sure it is not a header. The data is then extracted from the listview into an Order object and the completeStatus attribute is toggled from true to false or vise versa. From there it checks to see if the order still exists and if it does it will call DatabaseManager.changeOrderCompleteStatusValue() to change the value in the database.