Delete Functionality added to Orders Board
New pop up window in manager order board settings allows a user to select two radio buttons, one to delete a product and one to delete a customer. If the customer is selected, it will only show the drop down list of customers. To delete the customer the user must click the Confirm Delete button. If they want to delete a product, a second drop down window will appear with that customers products.

Validation is done to make sure that the correct inputs are required wether they are deleted a customer or a product. when pressing the delete button or the close window button, the delete button is disabled and the confirm delete radio button is also unchecked.

In OrderBoard Services, the functions deleteCustomerFromFarm() and deleteProductFromCustomerInFarm() are used to get the correct database reference to pass the customer name and if needed the product name to deleteCustomerFromFarm() and deleteProductFromCustomer() in the databaseManager class.