Shipping Calculator Added
Shipping calculator and database access and algorithm for calculating shipping added.

Also added On pause and resume in main activity that locks home screen and buttons and shows a loading symbol.

ShippingCalculator allows a user to select a product profile and a enter a quantity. It will tell them how to best pack pallets for shipping based off the parameters in the profile set by the manager.

Manager can use the class  ShippingCalculatorManagerProfiles to create, update and delete product profiles. All profiles are stored on the database and can be seen by all staff. Also has realtime updates for staff with the ShippingCalculator open