Added Leave farm option next to sign out
LEave farm removes the farm name from the user in the user document and in the usersInFarm document it sets currentlyInFarm to false. This allows a user to leave and rejoin a farm and keep their timetable.

A check is done when retrieving users for time table feature to check if the users are currently in the farm or not.

When a user joins a farm the write will also set their currentlyInFarm to true.