Timetable added to home page
Timetable use has been added to the home page and uses a modified version of the function used to update the timetable in the timetable activities. the timetable is pre loaded as a blank timetable with just the times and then updateMainMenuTimeTableWithTodaysDate is called which gets all the users tasks and sets them to the mini timetable. It lastly finds the current hour and sets that as the selected hour so the user can immediately see what they have coming up.