Main rework and tool tidying
Main was updated with a function to check if the user is the manager or not which sets a boolean to true if they are. It allows for two seperate functions attached to orders board and timetable buttons. It also does this check on load meaning that it opens those activities quicker.

Layouts in the tools were also updated and some code was tidied up in each tool class and the services class.