UI updates and Launcher Page and settings in create and join farm.
A new launcher page is the starting activity that checks if where the user should start. It checks if they are logged in or not and if they are in a farm bringing them to the login activity, the join farm activity or the main menu.

Settings allow users to logout at the join farm or create farm activities.